# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Small demo appliction serving albums",
        "title": "Kopwerk Demo",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "temp.swagger.io",
    "basePath": "/",
    "paths": {
        "/album/{id}": {
            "get": {
                "description": "get a album by its id",
                "produces": [
                    "application/json"
                ],
                "operationId": "get album by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/albums": {
            "get": {
                "description": "get all albums",
                "produces": [
                    "application/json"
                ],
                "operationId": "get albums",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/health": {
            "get": {
                "description": "get healthcheck",
                "produces": [
                    "application/json"
                ],
                "summary": "Check the health of the server",
                "operationId": "health",
                "responses": {}
            }
        }
    }
}